wip: refactor to use async call
